pyproject.toml
[project]
name = "freelacenter"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4",
    "requests",
    "google-generativeai",
    "schedule",
    "selenium",
    "python-dotenv",
]
